requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
rich==13.6.0
typing-extensions==4.8.0
requests==2.31.0
aiohttp==3.8.5 